techshop/web/requirements.txt
Django==4.2.11
mysqlclient==2.2.4
requests==2.31.0
djangorestframework==3.15.1
